+++ client/Packages/com.beamable/Common/Runtime/BeamCli/Contracts/GenerateEnvOutput.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+
+namespace Beamable.Common.BeamCli.Contracts
+{
+	[Serializable]
+	public class GenerateEnvFileOutput
+	{
+		public List<EnvVarOutput> envVars = new List<EnvVarOutput>();
+	}
+
+	[Serializable]
+	public class EnvVarOutput
+	{
+		public string name;
+		public string value;
+		public static EnvVarOutput Create(string name, string value) => new EnvVarOutput { name = name, value = value };
+	}
+}