+++ client/Packages/com.beamable/Editor/Modules/Content/ContentIO.cs
@@ -983,15 +983,12 @@ namespace Beamable.Editor.Content
 			content.SetContentName(nextName);
 			NotifyRenamed(oldId, content, nextAssetpath);
 			content.BroadcastUpdate();
+			
 			var result = AssetDatabase.MoveAsset(existingAssetPath, nextAssetpath);
 			if (!string.IsNullOrEmpty(result))
 			{
 				throw new Exception(result);
 			}
-
-			EditorUtility.SetDirty(content);
-			AssetDatabase.ForceReserializeAssets(new[] { nextAssetpath },
-				ForceReserializeAssetsOptions.ReserializeAssetsAndMetadata);
 		}
 
 
@@ -1169,7 +1166,12 @@ namespace Beamable.Editor.Content
 
 				var version = serverReference.version;
 				content.SetIdAndVersion(content.Id, version);
-				contentData[i] = new ContentDataInfo { contentId = content.Id, data = content.ToJson() };
+				contentData[i] = new ContentDataInfo
+				{
+					contentId = content.Id, 
+					contentVersion = content.Version,
+					data = content.ToJson()
+				};
 			}
 
 			ContentDataInfoWrapper fileData = new ContentDataInfoWrapper { content = contentData.ToList() };