+++ client/Packages/com.beamable/CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.17]
+### Changed
+- Baked content exists in a separate memory location than the deserialized `content.json` CDN cache
+
+### Fixed
+- Content cache evicts old content for new content id versions
+- `PlatformRequester` will reattempt failed requests caused by SSL connection issues.
+- Fix renaming content throwing infinite warnings
+
 ## [1.19.16]
 ### Fixed
 - Disabling Content inspectors no longer causes compiler errors.
 - Leaderboard `rankgt` field is not null when specifying outlier.
-- Fix renaming content throwing infinite warnings
 
 ## [1.19.15]
 ### Changed